Update Day to receive raw input instead of lines
Although it is often useful to split the puzzle input into lines, that
is not always the right choices.

A good example is 2022 day 1, where the input lines are grouped with an
empty line between groups. Splitting the raw input on a *double* line
break first is easier than trying to partition a list after it's already
been split on a single line break.